fix: move RemoveMethod visitor to MethodVisitors namespace
